Refine Notify Timing
Since we don't upgrade packages, there are no feasible scenario that
requires restarting webserver when installing packages (restart in
normal usage is covered by `configuration.yml`).  However, we might need
to restart webserver if the database is upgraded (which might occur when
someone had a running environment and upgraded the package manually but
didn't run the migration).

Signed-off-by: Ryo Tagami <rtagami@airstrip.jp>